+++ lib/additionals.rb
@@ -74,9 +74,9 @@ if ActiveRecord::Base.connection.table_exists?(:settings)
     require_dependency 'additionals/hooks'
 
     # Macros
-    Additionals.load_macros(%w[calendar cryptocompare date gist issue last_updated_at last_updated_by
-                               member project recently_updated reddit slideshare tradingview
-                               twitter user vimeo youtube])
+    Additionals.load_macros(%w[calendar cryptocompare date gist group_users issue last_updated_at
+                               last_updated_by member project recently_updated reddit slideshare
+                               tradingview twitter user vimeo youtube])
   end
 
   # include deface overwrites